refactor: move lax_numpy tensor contractions into their own file
