+++ jax/numpy/__init__.py
@@ -73,7 +73,6 @@ from jax._src.numpy.lax_numpy import (
     diagonal as diagonal,
     diff as diff,
     digitize as digitize,
-    dot as dot,
     dsplit as dsplit,
     dstack as dstack,
     dtype as dtype,
@@ -115,7 +114,6 @@ from jax._src.numpy.lax_numpy import (
     iinfo as iinfo,
     indices as indices,
     inf as inf,
-    inner as inner,
     insert as insert,
     interp as interp,
     isclose as isclose,
@@ -133,8 +131,6 @@ from jax._src.numpy.lax_numpy import (
     load as load,
     logspace as logspace,
     mask_indices as mask_indices,
-    matmul as matmul,
-    matvec as matvec,
     matrix_transpose as matrix_transpose,
     meshgrid as meshgrid,
     moveaxis as moveaxis,
@@ -145,7 +141,6 @@ from jax._src.numpy.lax_numpy import (
     ndim as ndim,
     newaxis as newaxis,
     nonzero as nonzero,
-    outer as outer,
     packbits as packbits,
     pad as pad,
     permute_dims as permute_dims,
@@ -179,7 +174,6 @@ from jax._src.numpy.lax_numpy import (
     swapaxes as swapaxes,
     take as take,
     take_along_axis as take_along_axis,
-    tensordot as tensordot,
     tile as tile,
     trace as trace,
     trapezoid as trapezoid,
@@ -198,9 +192,6 @@ from jax._src.numpy.lax_numpy import (
     unstack as unstack,
     unwrap as unwrap,
     vander as vander,
-    vdot as vdot,
-    vecdot as vecdot,
-    vecmat as vecmat,
     vsplit as vsplit,
     vstack as vstack,
     where as where,
@@ -258,6 +249,18 @@ from jax._src.numpy.sorting import (
     partition as partition,
     sort as sort,
     sort_complex as sort_complex,
+)
+
+from jax._src.numpy.tensor_contractions import (
+  dot as dot,
+  inner as inner,
+  matmul as matmul,
+  matvec as matvec,
+  outer as outer,
+  tensordot as tensordot,
+  vecdot as vecdot,
+  vecmat as vecmat,
+  vdot as vdot,
 )
 
 from jax._src.numpy.window_functions import (
